refactor(acp): collapse SDK and lazy runtime shims
