+++ src/auto-reply/reply/dispatch-acp-manager.runtime.ts
@@ -1,3 +1,4 @@
 /** Runtime ACP manager dependencies used by reply dispatch. */
 export { getAcpSessionManager } from "../../acp/control-plane/manager.js";
+export { readAcpSessionEntry } from "../../acp/runtime/session-meta.js";
 export { getSessionBindingService } from "../../infra/outbound/session-binding-service.js";
