--- src/auto-reply/reply/dispatch-acp-session.runtime.ts
@@ -1,2 +1,0 @@
-/** Runtime facade for reading persisted ACP session metadata. */
-export { readAcpSessionEntry } from "../../acp/runtime/session-meta.js";
