--- src/auto-reply/reply/dispatch-acp-tts.runtime.ts
@@ -1,2 +1,0 @@
-/** Runtime facade for applying TTS metadata to ACP reply payloads. */
-export { maybeApplyTtsToPayload } from "../../tts/tts.runtime.js";
