--- src/channels/plugins/acp-stateful-target-reset.runtime.ts
@@ -1,3 +1,0 @@
-// Runtime facade for ACP stateful target reset. Kept isolated so binding helpers can depend on
-// the reset seam without importing broader gateway runtime modules.
-export { performGatewaySessionReset } from "../../gateway/session-reset-service.js";
