+++ src/plugin-sdk/acp-binding-resolve-runtime.ts
@@ -1,2 +1,3 @@
-// Read-only ACP binding resolution without the lifecycle runtime import graph.
+// Deprecated compatibility for the released Matrix hot-path import.
+// Remove next major after supported Matrix versions resolve bindings without this host facade.
 export { resolveConfiguredAcpBindingRecord } from "../acp/persistent-bindings.resolve.js";
