+++ src/plugin-sdk/acp-binding-runtime.ts
@@ -1,5 +1,5 @@
-// Narrow ACP binding helpers for plugins that need persistent ACP setup state
-// without importing the broad core SDK surface.
+// Deprecated compatibility for released @openclaw/matrix packages.
+// Remove next major after supported Matrix versions no longer import this lazy lifecycle surface.
 
 export { ensureConfiguredAcpBindingReady } from "../acp/persistent-bindings.lifecycle.js";
 export { resolveConfiguredAcpBindingRecord } from "../acp/persistent-bindings.resolve.js";
